@@ -162,14 +162,14 @@ public class NettyMqttConverter {
     }
 
     private static PublishMsg extractPublishMsg(MqttPublishMessage mqttPublishMessage) {
-        byte[] payloadBytes = toBytes(mqttPublishMessage.payload());
+        ByteBuf byteBuf = mqttPublishMessage.payload().retain();
         return PublishMsg.builder()
                 .packetId(mqttPublishMessage.variableHeader().packetId())
                 .topicName(mqttPublishMessage.variableHeader().topicName())
                 .qosLevel(mqttPublishMessage.fixedHeader().qosLevel().value())
                 .isRetained(mqttPublishMessage.fixedHeader().isRetain())
                 .isDup(mqttPublishMessage.fixedHeader().isDup())
-                .payload(payloadBytes)
+                .byteBuf(byteBuf)
                 .properties(mqttPublishMessage.variableHeader().properties())
                 .build();
     }
@@ -185,7 +185,7 @@ public class NettyMqttConverter {
                 .build();
     }
 
-    private static byte[] toBytes(ByteBuf inbound) {
+    public static byte[] toBytes(ByteBuf inbound) {
         byte[] bytes = new byte[inbound.readableBytes()];
         int readerIndex = inbound.readerIndex();
         inbound.getBytes(readerIndex, bytes);
