@@ -17,13 +17,23 @@ package org.thingsboard.mqtt.broker.config;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.thingsboard.mqtt.broker.util.MiscUtils;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
 
 @Controller
 public class WebConfig {
 
-    @RequestMapping(value = {"/{path:^(?!api$)(?!assets$)(?!static$)(?!webjars$)[^\\.]*}/**"})
+    @RequestMapping(value = {"/{path:^(?!api$)(?!assets$)(?!static$)(?!webjars$)(?!swagger-ui$)[^\\.]*}/**"})
     public String redirect() {
         return "forward:/index.html";
     }
 
+    @RequestMapping("/swagger-ui.html")
+    public void redirectSwagger(HttpServletRequest request, HttpServletResponse response) throws IOException {
+        String baseUrl = MiscUtils.constructBaseUrl(request);
+        response.sendRedirect(baseUrl + "/swagger-ui/");
+    }
 }
