@@ -15,7 +15,6 @@
  */
 package org.thingsboard.mqtt.broker.service.mqtt.retain;
 
-import io.netty.handler.codec.mqtt.MqttProperties;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.Before;
 import org.junit.Test;
@@ -70,6 +69,6 @@ public class RetainedMsgProcessorImplTest {
     }
 
     private PublishMsg getMsg(byte[] payload) {
-        return new PublishMsg(1, "test", payload, 2, true, false, MqttProperties.NO_PROPERTIES);
+        return new PublishMsg(1, "test", payload, 2, true, false);
     }
 }
